cli_commands: Moved everything probe related to their own module and restructured everything to support it.
